refactor(import): extract three-phase Bambu upsert into src/lib/bambuUpsert.ts
Move the battle-hardened active → resurrect-trashed → create-with-E11000-race
upsert (plus augmentExistingWithParent / composeMongoUpdate) out of the
bambustudio route into a shared lib, so the upcoming bulk OrcaSlicer library
importer can run the exact same write path per record without the two routes
drifting (same rationale as iniImportApply.ts on the INI side).

upsertParsedBambuFilament returns a result object instead of NextResponses;
expected failures (payload validation, missing type/vendor, Mongoose
validator rejections) map to { ok: false, status, error } and the route
re-wraps them, so response shapes are unchanged. New: an optional parentId
applied ONLY on the phase-3 create — update phases never re-parent an
existing row.
